+++ windows/flutter/generated_plugins.cmake
@@ -6,6 +6,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   app_links
   audioplayers_windows
   connectivity_plus
+  flutter_webrtc
   geolocator_windows
   permission_handler_windows
   record_windows